pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llama-vllm"
version = "0.1.0"
description = "A comprehensive framework for model distillation, fine-tuning, and inference using vLLM and LLaMA Factory"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "LLaMA-vLLM Team" }]
keywords = ["llm", "distillation", "fine-tuning", "inference", "vllm", "llama"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.3.0",
    "transformers>=4.41.0",
    "accelerate>=0.30.0",
    "datasets>=2.20.0",
    "tokenizers>=0.19.0",
    "pydantic>=2.7.0",
    "pyyaml>=6.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "tqdm>=4.66.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "sentencepiece>=0.2.0",
    "einops>=0.8.0",
]

[project.optional-dependencies]
train = [
    "peft>=0.11.0",
    "trl>=0.9.0",
    "bitsandbytes>=0.43.0",
    "deepspeed>=0.14.0",
    "wandb>=0.17.0",
    "tensorboard>=2.17.0",
    "tiktoken>=0.7.0",
]
inference = [
    "vllm>=0.5.0",
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "sse-starlette>=2.1.0",
    "httpx>=0.27.0",
]
quant = [
    "auto-gptq>=0.7.0",
    "autoawq>=0.2.5",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
    "isort>=5.13.0",
]

[project.scripts]
llama-vllm = "llama_vllm.cli.main:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]

